64-bit fixes for printf() arguments.
darcs-hash:20060925222351-65a35-8ad906440a4f9b1e6f6979a6b8491de1750917d0.gz

+++ libsofia-sip-ua/nua/outbound.c
@@ -294,7 +294,7 @@ int outbound_set_options(outbound_t *ob,
     else if (MATCH(use-upnp) || MATCH(use_upnp)) prefs->use_upnp = value;
     else if (MATCH(use-stun) || MATCH(use_stun)) prefs->use_stun = value;
     else
-      SU_DEBUG_1(("outbound_t: unknown option \"%.*s\"\n", len, s));
+      SU_DEBUG_1(("outbound_t: unknown option \"%.*s\"\n", (int)len, s));
 
     s += len;
     len = strspn(s, " \t\n\r,;");
